requirements.txt
tensorflow-cpu==2.13.0
streamlit==1.28.0
opencv-python-headless==4.8.1
numpy==1.24.3
Pillow==10.0.1